crud de autores

+++ microservicios/canciones/src/main/java/com/infraestructura/canciones/Repository/CancionRepository.java
@@ -8,7 +8,7 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CancionRepository extends CrudRepository<Cancion, Integer>{
+public interface CancionRepository extends CrudRepository<Cancion, Long>{
     Optional<Cancion> findById(Long id);
     void deleteById(Long id);
     Optional<Cancion> findByTitulo(String titulo);
